// display-tables/src/interfaces/consumption.ts
export interface ConsumptionInterface {
  totalConsumptionDay: number;
  totalConsumptionNight: number;
  minerConsumptionDay: number;
  minerConsumptionNight: number;
  housingWorkerConsumptionDay: number;
  housingWorkerConsumptionNight: number;
  orelConsumptionDay: number;
  orelConsumptionNight: number;
}